v2.5.0 OtpVerification.Rpc

This module contains functions that are called from other pods via RPC.

Types

phone_response()

phone_response() :: %{phone_number: binary()}

sms()

sms() :: %{body: binary(), id: binary(), phone_number: binary(), type: binary()}

verification()

verification() :: %{
  active: boolean(),
  code_expired_at: DateTime,
  id: binary(),
  status: binary()
}

Functions

complete(phone_number, code)

complete(phone_number :: binary(), code :: integer()) ::
  {:ok, verification()} | {:error, atom()} | nil

Complete code by phone_number

Examples

iex> OtpVerification.Rpc.complete("+380960000000", 123456)
{:ok,
%{
  active: false,
  code_expired_at: #DateTime<2019-08-07 00:00:00.000000Z>,
  id: "448c7356-1b69-4f4f-be70-d7e558b7cb09",
  status: "verified"
}}

initialize(phone_number, provider \\ nil)

initialize(phone_number :: binary(), provider :: binary() | nil) ::
  {:ok, verification()} | {:error, atom()}

Initialize sms by phone number and provider

Examples

iex> OtpVerification.Rpc.initialize("+380960000000")
{:ok,
%{
  active: true,
  code_expired_at: #DateTime<2019-05-02 21:58:51.902086Z>,
  id: "53982ced-ed95-48a9-bab7-78247ec7259f",
  status: "new"
}}

send_sms(phone_number, body, type, provider \\ nil)

send_sms(
  phone_number :: binary(),
  body :: binary(),
  type :: binary(),
  provider :: binary()
) :: {:ok, sms()} | {:error, any()}

Send sms to phone number with body, type and provider

Examples

iex> OtpVerification.Rpc.send_sms("+380960000000", "foo", "undefined")
{:ok,
%{
  body: "foo",
  id: "08600dea-f366-44e4-ad44-7111368f2209",
  phone_number: "+380960000000",
  type: "undefined"
}}

verification_phone(phone_number)

verification_phone(phone_number :: binary()) :: {:ok, phone_response()} | nil

Get verification phone by phone number

Examples

iex> OtpVerification.Rpc.verification_phone("+380960000000")
{:ok, %{phone_number: "+380960000000"}}